Add the LACommenter plugin
The LACcommenter will comment literal arguments in function calls. The
following argument types are commented:
  * integer
  * floating point
  * character
  * boolean
  * string
The injected comments adhere to LLVM's coding guideline:
  * https://llvm.org/docs/CodingStandards.html#comment-formatting

This plugin is based on an example presented by Peter Smith at Linaro Connect
2018:
  * https://s3.amazonaws.com/connect.linaro.org/yvr18/presentations/yvr18-223.pdf
This is a fresh implementation of the ideas presented by Peter,
complimented with CMake scripts and tests.