[GA] Use Clang when building Clang/LLVM from sources
Ubuntu 18.04 provides gcc-7 as the default compiler. Clang (and LLVM)
build just fine with that, but then some clang-tutor plugins fail with
an internal Clang error. I'm not quite sure why.  Everything is fine
when building with system clang (the available version is clang-9). So,
instead of debugging the above, I'm using clang to build Clang/LLVM for
clang-tutor.