Fix nullptr dereference in Obfuscator
When there are no `ImplicitCastExpr` (e.g. when arguments are literal),
the following returns `nullptr`:
```
Result.Nodes.getNodeAs<clang::ImplicitCastExpr>("lhs")
```

This means that the following leads to a seg fault:
```
Result.Nodes.getNodeAs<clang::ImplicitCastExpr>("lhs")->IgnoreImplicit()
```

I've managed to avoid seg-faults for a while, but I think that I was simply lucky
with the compilers that I was using. This patch fixes that.

A seg-fault can be reproduced with a Release build of Clang. Make sure
to build it with GCC. I tested this with GCC 7.4 and GCC 9.2. Building
Clang with Clang meant that there were no seg-faults. Use
test/MBA_add_int.cpp to reproduce.

This patch also removes the necessity for the workaround added in:
e806a1ebea51df5e8b10e5e14e118fa6aed0f786