Fix Ubuntu configuration
The `apt` packages for Ubuntu have recently been updated to include
MLIR. This has triggered an update in the LLVM CMake modules and
`libmlir-13` is now required to be installed alongside other LLVM
packaged (otherwise CMake configuration fails). I've added a couple more
LLVM packages (`libmlir-13-dev` and `libllvm13`), just to be on the safe
side.

To make things a bit easier for myself, I've switched from Ubuntu 18.04
to 20.04. I think that this is rather safe given that 20.04 is almost 2
years old.